Refactors the server splitting it
